moved plugin specific types into concrete plugin package
