+++ plugin/dca/dca_test.go
@@ -18,6 +18,8 @@ import (
 	"github.com/vultisig/mobile-tss-lib/tss"
 	"github.com/vultisig/vultisigner/common"
 	"github.com/vultisig/vultisigner/internal/types"
+	pg "github.com/vultisig/vultisigner/plugin"
+
 	"github.com/vultisig/vultisigner/test/mocks/database"
 	"github.com/vultisig/vultisigner/test/mocks/ethclient"
 	"github.com/vultisig/vultisigner/test/mocks/uniswapclient"
@@ -33,13 +35,13 @@ func createSamplePolicy(progress string) types.PluginPolicy {
 }
 
 func createValidPolicy() types.PluginPolicy {
-	dcaPolicy := types.DCAPolicy{
+	dcaPolicy := DCAPolicy{
 		ChainID:            "1",
 		SourceTokenID:      "0x1111111111111111111111111111111111111111",
 		DestinationTokenID: "0x2222222222222222222222222222222222222222",
 		TotalAmount:        "1000",
 		TotalOrders:        "10",
-		Schedule: types.Schedule{
+		Schedule: pg.Schedule{
 			Frequency: "daily",
 			Interval:  "1",
 		},
@@ -246,13 +248,13 @@ func TestValidatePluginPolicy(t *testing.T) {
 			name: "Invalid source and destination tokens",
 			policy: func() types.PluginPolicy {
 				p := createValidPolicy()
-				dcaPolicy := types.DCAPolicy{
+				dcaPolicy := DCAPolicy{
 					ChainID:            "1",
 					SourceTokenID:      "0x1111111111111111111111111111111111111111",
 					DestinationTokenID: "0x1111111111111111111111111111111111111111", // Same as source
 					TotalAmount:        "1000",
 					TotalOrders:        "10",
-					Schedule: types.Schedule{
+					Schedule: pg.Schedule{
 						Frequency: "daily",
 						Interval:  "1",
 					},
@@ -267,13 +269,13 @@ func TestValidatePluginPolicy(t *testing.T) {
 			name: "Invalid total amount",
 			policy: func() types.PluginPolicy {
 				p := createValidPolicy()
-				dcaPolicy := types.DCAPolicy{
+				dcaPolicy := DCAPolicy{
 					ChainID:            "1",
 					SourceTokenID:      "0x1111111111111111111111111111111111111111",
 					DestinationTokenID: "0x2222222222222222222222222222222222222222",
 					TotalAmount:        "-1",
 					TotalOrders:        "10",
-					Schedule: types.Schedule{
+					Schedule: pg.Schedule{
 						Frequency: "daily",
 						Interval:  "1",
 					},
@@ -288,13 +290,13 @@ func TestValidatePluginPolicy(t *testing.T) {
 			name: "Invalid total orders",
 			policy: func() types.PluginPolicy {
 				p := createValidPolicy()
-				dcaPolicy := types.DCAPolicy{
+				dcaPolicy := DCAPolicy{
 					ChainID:            "1",
 					SourceTokenID:      "0x1111111111111111111111111111111111111111",
 					DestinationTokenID: "0x2222222222222222222222222222222222222222",
 					TotalAmount:        "1000",
 					TotalOrders:        "0",
-					Schedule: types.Schedule{
+					Schedule: pg.Schedule{
 						Frequency: "daily",
 						Interval:  "1",
 					},
@@ -309,17 +311,17 @@ func TestValidatePluginPolicy(t *testing.T) {
 			name: "Invalid price range",
 			policy: func() types.PluginPolicy {
 				p := createValidPolicy()
-				dcaPolicy := types.DCAPolicy{
+				dcaPolicy := DCAPolicy{
 					ChainID:            "1",
 					SourceTokenID:      "0x1111111111111111111111111111111111111111",
 					DestinationTokenID: "0x2222222222222222222222222222222222222222",
 					TotalAmount:        "1000",
 					TotalOrders:        "10",
-					PriceRange: types.PriceRange{
+					PriceRange: PriceRange{
 						Min: "200",
 						Max: "100", // Min > Max
 					},
-					Schedule: types.Schedule{
+					Schedule: pg.Schedule{
 						Frequency: "daily",
 						Interval:  "1",
 					},
@@ -344,13 +346,13 @@ func TestValidatePluginPolicy(t *testing.T) {
 			name: "Invalid frequency",
 			policy: func() types.PluginPolicy {
 				p := createValidPolicy()
-				dcaPolicy := types.DCAPolicy{
+				dcaPolicy := DCAPolicy{
 					ChainID:            "1",
 					SourceTokenID:      "0x1111111111111111111111111111111111111111",
 					DestinationTokenID: "0x2222222222222222222222222222222222222222",
 					TotalAmount:        "1000",
 					TotalOrders:        "10",
-					Schedule: types.Schedule{
+					Schedule: pg.Schedule{
 						Frequency: "invalid",
 						Interval:  "1",
 					},
@@ -534,7 +536,7 @@ func TestProposeTransactions(t *testing.T) {
 			name:   "All swaps completed",
 			policy: policy,
 			mockSetup: func(db *database.MockDB, mockUniswap *uniswapclient.MockUniswapClient) {
-				dcaPolicy := types.DCAPolicy{}
+				dcaPolicy := DCAPolicy{}
 				err := json.Unmarshal(policy.Policy, &dcaPolicy)
 				require.NoError(t, err)
 
@@ -820,7 +822,7 @@ func TestValidateProposedTransactions(t *testing.T) {
 			policy:     policy,
 			txRequests: []types.PluginKeysignRequest{validSwapRequest},
 			mockSetup: func(db *database.MockDB, uniswap *uniswapclient.MockUniswapClient) {
-				dcaPolicy := types.DCAPolicy{}
+				dcaPolicy := DCAPolicy{}
 				err := json.Unmarshal(policy.Policy, &dcaPolicy)
 				require.NoError(t, err)
 				totalOrders, _ := new(big.Int).SetString(dcaPolicy.TotalOrders, 10)
